[Feat]: Usecase 기능 수정
